TST: GH26807 separate out test_categorical
