+++ pandas/tests/io/pytables/test_store.py
@@ -10,7 +10,6 @@ import pytest
 
 import pandas as pd
 from pandas import (
-    Categorical,
     DataFrame,
     DatetimeIndex,
     Index,
@@ -32,15 +31,7 @@ from pandas.tests.io.pytables.common import (
     tables,
 )
 
-from pandas.io.pytables import (
-    ClosedFileError,
-    HDFStore,
-    Term,
-    _maybe_adjust_name,
-    read_hdf,
-)
-
-from pandas.io import pytables as pytables  # isort:skip
+from pandas.io.pytables import HDFStore, Term, _maybe_adjust_name, read_hdf
 
 _default_compressor = "blosc"
 ignore_natural_naming_warning = pytest.mark.filterwarnings(
@@ -1795,133 +1786,6 @@ class TestHDFStore:
         result = tm.round_trip_localpath(writer, reader)
         tm.assert_frame_equal(df, result)
 
-    def test_multiple_open_close(self, setup_path):
-        # gh-4409: open & close multiple times
-
-        with ensure_clean_path(setup_path) as path:
-
-            df = tm.makeDataFrame()
-            df.to_hdf(path, "df", mode="w", format="table")
-
-            # single
-            store = HDFStore(path)
-            assert "CLOSED" not in store.info()
-            assert store.is_open
-
-            store.close()
-            assert "CLOSED" in store.info()
-            assert not store.is_open
-
-        with ensure_clean_path(setup_path) as path:
-
-            if pytables._table_file_open_policy_is_strict:
-                # multiples
-                store1 = HDFStore(path)
-                msg = (
-                    r"The file [\S]* is already opened\.  Please close it before "
-                    r"reopening in write mode\."
-                )
-                with pytest.raises(ValueError, match=msg):
-                    HDFStore(path)
-
-                store1.close()
-            else:
-
-                # multiples
-                store1 = HDFStore(path)
-                store2 = HDFStore(path)
-
-                assert "CLOSED" not in store1.info()
-                assert "CLOSED" not in store2.info()
-                assert store1.is_open
-                assert store2.is_open
-
-                store1.close()
-                assert "CLOSED" in store1.info()
-                assert not store1.is_open
-                assert "CLOSED" not in store2.info()
-                assert store2.is_open
-
-                store2.close()
-                assert "CLOSED" in store1.info()
-                assert "CLOSED" in store2.info()
-                assert not store1.is_open
-                assert not store2.is_open
-
-                # nested close
-                store = HDFStore(path, mode="w")
-                store.append("df", df)
-
-                store2 = HDFStore(path)
-                store2.append("df2", df)
-                store2.close()
-                assert "CLOSED" in store2.info()
-                assert not store2.is_open
-
-                store.close()
-                assert "CLOSED" in store.info()
-                assert not store.is_open
-
-                # double closing
-                store = HDFStore(path, mode="w")
-                store.append("df", df)
-
-                store2 = HDFStore(path)
-                store.close()
-                assert "CLOSED" in store.info()
-                assert not store.is_open
-
-                store2.close()
-                assert "CLOSED" in store2.info()
-                assert not store2.is_open
-
-        # ops on a closed store
-        with ensure_clean_path(setup_path) as path:
-
-            df = tm.makeDataFrame()
-            df.to_hdf(path, "df", mode="w", format="table")
-
-            store = HDFStore(path)
-            store.close()
-
-            msg = r"[\S]* file is not open!"
-            with pytest.raises(ClosedFileError, match=msg):
-                store.keys()
-
-            with pytest.raises(ClosedFileError, match=msg):
-                "df" in store
-
-            with pytest.raises(ClosedFileError, match=msg):
-                len(store)
-
-            with pytest.raises(ClosedFileError, match=msg):
-                store["df"]
-
-            with pytest.raises(ClosedFileError, match=msg):
-                store.select("df")
-
-            with pytest.raises(ClosedFileError, match=msg):
-                store.get("df")
-
-            with pytest.raises(ClosedFileError, match=msg):
-                store.append("df2", df)
-
-            with pytest.raises(ClosedFileError, match=msg):
-                store.put("df3", df)
-
-            with pytest.raises(ClosedFileError, match=msg):
-                store.get_storer("df2")
-
-            with pytest.raises(ClosedFileError, match=msg):
-                store.remove("df2")
-
-            with pytest.raises(ClosedFileError, match=msg):
-                store.select("df")
-
-            msg = "'HDFStore' object has no attribute 'df'"
-            with pytest.raises(AttributeError, match=msg):
-                store.df
-
     def test_copy(self, setup_path):
 
         with catch_warnings(record=True):
@@ -2048,180 +1912,6 @@ class TestHDFStore:
             store.append("test", df, format="table", data_columns=True)
             result = store.select("test", 'a = "test & test"')
         tm.assert_frame_equal(expected, result)
-
-    def test_categorical(self, setup_path):
-
-        with ensure_clean_store(setup_path) as store:
-
-            # Basic
-            _maybe_remove(store, "s")
-            s = Series(
-                Categorical(
-                    ["a", "b", "b", "a", "a", "c"],
-                    categories=["a", "b", "c", "d"],
-                    ordered=False,
-                )
-            )
-            store.append("s", s, format="table")
-            result = store.select("s")
-            tm.assert_series_equal(s, result)
-
-            _maybe_remove(store, "s_ordered")
-            s = Series(
-                Categorical(
-                    ["a", "b", "b", "a", "a", "c"],
-                    categories=["a", "b", "c", "d"],
-                    ordered=True,
-                )
-            )
-            store.append("s_ordered", s, format="table")
-            result = store.select("s_ordered")
-            tm.assert_series_equal(s, result)
-
-            _maybe_remove(store, "df")
-            df = DataFrame({"s": s, "vals": [1, 2, 3, 4, 5, 6]})
-            store.append("df", df, format="table")
-            result = store.select("df")
-            tm.assert_frame_equal(result, df)
-
-            # Dtypes
-            _maybe_remove(store, "si")
-            s = Series([1, 1, 2, 2, 3, 4, 5]).astype("category")
-            store.append("si", s)
-            result = store.select("si")
-            tm.assert_series_equal(result, s)
-
-            _maybe_remove(store, "si2")
-            s = Series([1, 1, np.nan, 2, 3, 4, 5]).astype("category")
-            store.append("si2", s)
-            result = store.select("si2")
-            tm.assert_series_equal(result, s)
-
-            # Multiple
-            _maybe_remove(store, "df2")
-            df2 = df.copy()
-            df2["s2"] = Series(list("abcdefg")).astype("category")
-            store.append("df2", df2)
-            result = store.select("df2")
-            tm.assert_frame_equal(result, df2)
-
-            # Make sure the metadata is OK
-            info = store.info()
-            assert "/df2   " in info
-            # assert '/df2/meta/values_block_0/meta' in info
-            assert "/df2/meta/values_block_1/meta" in info
-
-            # unordered
-            _maybe_remove(store, "s2")
-            s = Series(
-                Categorical(
-                    ["a", "b", "b", "a", "a", "c"],
-                    categories=["a", "b", "c", "d"],
-                    ordered=False,
-                )
-            )
-            store.append("s2", s, format="table")
-            result = store.select("s2")
-            tm.assert_series_equal(result, s)
-
-            # Query
-            _maybe_remove(store, "df3")
-            store.append("df3", df, data_columns=["s"])
-            expected = df[df.s.isin(["b", "c"])]
-            result = store.select("df3", where=['s in ["b","c"]'])
-            tm.assert_frame_equal(result, expected)
-
-            expected = df[df.s.isin(["b", "c"])]
-            result = store.select("df3", where=['s = ["b","c"]'])
-            tm.assert_frame_equal(result, expected)
-
-            expected = df[df.s.isin(["d"])]
-            result = store.select("df3", where=['s in ["d"]'])
-            tm.assert_frame_equal(result, expected)
-
-            expected = df[df.s.isin(["f"])]
-            result = store.select("df3", where=['s in ["f"]'])
-            tm.assert_frame_equal(result, expected)
-
-            # Appending with same categories is ok
-            store.append("df3", df)
-
-            df = concat([df, df])
-            expected = df[df.s.isin(["b", "c"])]
-            result = store.select("df3", where=['s in ["b","c"]'])
-            tm.assert_frame_equal(result, expected)
-
-            # Appending must have the same categories
-            df3 = df.copy()
-            df3["s"] = df3["s"].cat.remove_unused_categories()
-
-            msg = (
-                "cannot append a categorical with different categories to the existing"
-            )
-            with pytest.raises(ValueError, match=msg):
-                store.append("df3", df3)
-
-            # Remove, and make sure meta data is removed (its a recursive
-            # removal so should be).
-            result = store.select("df3/meta/s/meta")
-            assert result is not None
-            store.remove("df3")
-
-            with pytest.raises(
-                KeyError, match="'No object named df3/meta/s/meta in the file'"
-            ):
-                store.select("df3/meta/s/meta")
-
-    def test_categorical_conversion(self, setup_path):
-
-        # GH13322
-        # Check that read_hdf with categorical columns doesn't return rows if
-        # where criteria isn't met.
-        obsids = ["ESP_012345_6789", "ESP_987654_3210"]
-        imgids = ["APF00006np", "APF0001imm"]
-        data = [4.3, 9.8]
-
-        # Test without categories
-        df = DataFrame({"obsids": obsids, "imgids": imgids, "data": data})
-
-        # We are expecting an empty DataFrame matching types of df
-        expected = df.iloc[[], :]
-        with ensure_clean_path(setup_path) as path:
-            df.to_hdf(path, "df", format="table", data_columns=True)
-            result = read_hdf(path, "df", where="obsids=B")
-            tm.assert_frame_equal(result, expected)
-
-        # Test with categories
-        df.obsids = df.obsids.astype("category")
-        df.imgids = df.imgids.astype("category")
-
-        # We are expecting an empty DataFrame matching types of df
-        expected = df.iloc[[], :]
-        with ensure_clean_path(setup_path) as path:
-            df.to_hdf(path, "df", format="table", data_columns=True)
-            result = read_hdf(path, "df", where="obsids=B")
-            tm.assert_frame_equal(result, expected)
-
-    def test_categorical_nan_only_columns(self, setup_path):
-        # GH18413
-        # Check that read_hdf with categorical columns with NaN-only values can
-        # be read back.
-        df = DataFrame(
-            {
-                "a": ["a", "b", "c", np.nan],
-                "b": [np.nan, np.nan, np.nan, np.nan],
-                "c": [1, 2, 3, 4],
-                "d": Series([None] * 4, dtype=object),
-            }
-        )
-        df["a"] = df.a.astype("category")
-        df["b"] = df.b.astype("category")
-        df["d"] = df.b.astype("category")
-        expected = df
-        with ensure_clean_path(setup_path) as path:
-            df.to_hdf(path, "df", format="table", data_columns=True)
-            result = read_hdf(path, "df")
-            tm.assert_frame_equal(result, expected)
 
     def test_duplicate_column_name(self, setup_path):
         df = DataFrame(columns=["a", "a"], data=[[0, 0]])
